refactor(tui-react): make main entry point browser-safe for Storybook
Split Node.js-only imports (node:util, node:fs) out of the main '.' entry
point to fix Storybook/Vite builds that externalize Node built-ins.

- Replace node:util formatWithOptions with Effect's Inspectable.toStringUnknown
- Move Node-only detection functions (isPiped, isRedirectedToFile, etc.) to
  OutputMode.node.ts
- Add new '@overeng/tui-react/node' export path for Node-only consumers
- Update all consumers (genie, megarepo, notion-cli, examples, tests) to import
  Node-only symbols from the /node subpath
