+++ packages/@overeng/tui-react/examples/02-effect-integration/counter.tsx
@@ -21,7 +21,8 @@ import { NodeContext, NodeRuntime } from '@effect/platform-node'
 import { Duration, Effect } from 'effect'
 import React from 'react'
 
-import { createTuiApp, run, outputOption, outputModeLayer } from '../../src/mod.ts'
+import { createTuiApp, run } from '../../src/mod.ts'
+import { outputOption, outputModeLayer } from '../../src/node/mod.ts'
 // Import from shared modules
 import { CounterState, CounterAction, counterReducer } from './schema.ts'
 import { CounterView } from './view.tsx'
