+++ packages/@overeng/tui-react/examples/04-stress-tests/rapid-updates.tsx
@@ -19,7 +19,8 @@ import { NodeContext, NodeRuntime } from '@effect/platform-node'
 import { Effect, Fiber } from 'effect'
 import React from 'react'
 
-import { createTuiApp, run, outputOption, outputModeLayer } from '../../src/mod.ts'
+import { createTuiApp, run } from '../../src/mod.ts'
+import { outputOption, outputModeLayer } from '../../src/node/mod.ts'
 // Import from shared modules
 import { StressTestState, StressTestAction, createStressTestReducer } from './schema.ts'
 import { StressTestView } from './view.tsx'
