+++ packages/@overeng/tui-react/examples/05-advanced/bouncing-windows.tsx
@@ -22,7 +22,8 @@ import { NodeContext, NodeRuntime } from '@effect/platform-node'
 import { Effect, Fiber } from 'effect'
 import React from 'react'
 
-import { createTuiApp, run, outputOption, outputModeLayer } from '../../src/mod.ts'
+import { createTuiApp, run } from '../../src/mod.ts'
+import { outputOption, outputModeLayer } from '../../src/node/mod.ts'
 // Import from shared modules
 import { AppState, AppAction, appReducer, createWindow } from './schema.ts'
 import { BouncingWindowsView } from './view.tsx'
